friends/requirements/dev.txt
-r base.txt
flake8==7.3.0
isort==6.0.1
pre-commit==4.2.0
pytest
pytest-django
